finished

+++ src/main/java/com/dom/employeemanager/service/OtpCodeService.java
@@ -1,10 +1,15 @@
 package com.dom.employeemanager.service;
 
+import com.dom.employeemanager.dtos.ResetPasswordDTO;
+import com.dom.employeemanager.exception.UserNotFoundException;
+import com.dom.employeemanager.models.Employee;
 import com.dom.employeemanager.models.OtpCode;
+import com.dom.employeemanager.repo.EmployeeRepo;
 import com.dom.employeemanager.repo.OtpCodeRepository;
 import jakarta.mail.MessagingException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -15,32 +20,45 @@ import java.util.Random;
 @RequiredArgsConstructor
 public class OtpCodeService {
   private final OtpCodeRepository otpCodeRepository;
-
+  private final EmployeeRepo employeeRepo;
   private final EmailService emailService;
+  private final PasswordEncoder passwordEncoder;
 
   @Value("${otp.expiration.time}")
   private long otpExpirationTime;  // Đơn vị: phút
 
-  public void generateAndSendOtpCode(String email) throws MessagingException {
-    // Xoá OTP cũ nếu tồn tại
-    otpCodeRepository.deleteByEmail(email);
+  public String generateAndSendOtpCode(String email, String phone) throws MessagingException {
+    // check db
+    Optional<Employee> existingEmployee = employeeRepo.findByEmailAndPhone(email, phone);
 
-    // Tạo OTP mới
-    String otpCode = String.format("%06d", new Random().nextInt(999999));
-    LocalDateTime expirationTime = LocalDateTime.now().plusMinutes(otpExpirationTime);
+    if (existingEmployee.isPresent()) {
+      // Kiểm tra nếu OTP cũ có tồn tại trước khi xoá
+      Optional<OtpCode> existingOtp = otpCodeRepository.findByEmail(email);
+      if (existingOtp.isPresent()) {
+        // Xoá OTP cũ nếu có
+        otpCodeRepository.deleteByEmail(email);
+      }
 
-    // Lưu OTP vào database
-    OtpCode optCodeNew = OtpCode.builder()
-      .email(email)
-      .otpCode(otpCode)
-      .expirationTime(expirationTime)
-      .build();
+      // Tạo OTP mới
+      String otpCode = String.format("%06d", new Random().nextInt(999999));
+      LocalDateTime expirationTime = LocalDateTime.now().plusMinutes(otpExpirationTime);
 
-    otpCodeRepository.save(optCodeNew);
+      // Lưu OTP vào database
+      OtpCode optCodeNew = OtpCode.builder()
+        .email(email)
+        .otpCode(otpCode)
+        .expirationTime(expirationTime)
+        .build();
 
-    // Gửi OTP qua email
-    String message = "Your OTP code is: " + otpCode + ". It will expire in " + otpExpirationTime + " minutes.";
-    emailService.sendOtpMessage(email, "OTP Code", message);
+      otpCodeRepository.save(optCodeNew);
+
+      // Gửi OTP qua email
+      String message = "Your OTP code is: " + otpCode + ". It will expire in " + otpExpirationTime + " minutes.";
+      emailService.sendOtpMessage(email, "OTP Code", message);
+      return otpCode;
+    } else {
+      throw new UserNotFoundException("User not existing");
+    }
   }
 
   public boolean validateOtp(String email, String otpCode) {
@@ -53,5 +71,25 @@ public class OtpCodeService {
       }
     }
     return false;  // OTP không hợp lệ
+  }
+
+  public boolean resetPassword(ResetPasswordDTO resetPasswordDTO) {
+    Optional<OtpCode> emailAndOtpCode =
+      otpCodeRepository.findByEmailAndOtpCode(resetPasswordDTO.getEmail(), resetPasswordDTO.getOtp());
+
+    // From otp table, get Email
+    if (emailAndOtpCode.isPresent()){
+      String existingEmail = emailAndOtpCode.get().getEmail();
+      Optional<Employee> existingEmployee = employeeRepo.findByEmail(existingEmail);
+      existingEmployee.ifPresent(employee -> {
+        // Set the new password
+        employee.setPassword(passwordEncoder.encode(resetPasswordDTO.getNewPassword()));
+
+        // Save the updated employee in the database
+        employeeRepo.save(employee);
+      });
+      return true;
+    }
+    return false;
   }
 }
